Rework make all
Previously make all would deplend on each of the three tarballs that we
wish to build. It would have rules that would use recursive make call
to build them. This is not what the current build system does (it simply
builds all three at the same time) which allows non-recursive android make
to share dependencies properly and to build things in proper order.

With this change that same logic is restored. The 'all' goal simply builds
the three tarballs with a recursive make call. Since the flash goal also
depends on the tarballs to create a bootable SD card a new dummy rule
was added that informs the user about the need to manually run
make all before make flash is available