Ignore empty values in pass-to-make
Some variables that are on pass-to-make may end up being empty. In that
case we'd pass them to make anyway as "make VAR=". That would be different
than expected as those variables would be defined empty and would not
take their expected default values.

This is a practical problem for TARGET_BUILD_VARIANT that defaults to 'eng'

To fix this we simply filter out all variables that expand to nothing