Fix make shell PS1
This fix is subtle and depends on the way we spawn commands for make shell.
Instead of using the default make's shell of /bin/sh, for this special
target we use /bin/bash -i

This ensures that all the right setup is made and that PS1 is configured
the way the user would expect.